update check file for automata test
Note: due to the way codegen works right now (iterating over g.nodes
in reverse order instead of proper SCC), we're using different symbols
for a function from inside vs. outside. This prevents cross
inside/outside CSE for nodes that depend on the function symbol, so
the code contains to occurences of

  new scala.lms.tutorial.Automaton[Char,Boolean](false, x15)

and both

  new scala.lms.tutorial.Automaton[Char,Boolean](false, x2)
  new scala.lms.tutorial.Automaton[Char,Boolean](false, x3)

which is not strictly necessary (also see note in lms-clean/traverser).